feat: bootstrap system w/ sandbox discoveries
This version does not contain all of the configurations from the
previous version I built, but it contains a large chunk of the config.
I used GNOME to bootstrap the system as it comes with nice configuration
tools out of the box (setting up wifi can be a bear without GUIs). The
next step is migrating over to something like Hyprland.