fix: update hyprland
Something in the last hyprland version did not agree with the rest of my
system. I thought it was a issue with my GPU firmware disagreeing with
the WM, but, no matter what changes I tried, I couldn't get hyprland to
spin up. Updating it again, though, seems to have fixed it. Looks like
it was just something wrong with that previous update?